Bootstrap Function Modification

This Bootstrap function will provide you always an object of class bootstrap.

Bootstrap(x, B = 1000, statistic, sample_size = length(x))

Arguments

x

represent the original sample

B

number of bootstrap samples. By default B = 1000

statistic

is an indicator of the statistic applied to the samples

sample_size

is the length of every bootstrap sample. By default sample_size = "the original sample size"

Value

A list of type bootstrap.class composed of relevant values provided by the Bootstrap function:

theta_hat

The statistic applied to the original sample

theta_star

A list of length B composed of statistics applied to each bootstrapped samples

varBoot

The variance applied to theta_star

Unbiased_estimator

Unbieased variance of the estimator

coerced_to_a_vector

Returning a True or Flase regarding the initial sample if it was coerced to a vector

stat

Return the name of the statistic used thruough the Bootstrap function

B

Number of bootrsapped samples

sample_size

Return the length of every bootstrap sample

Author

Group-E

Examples

Bootstrap(rnorm(100), statistic = mean)
#> Non-parametric bootstrap estimate of the variance of 'mean'
#> -----------------
#> Variance estimate is 0.009 
#> Based on B=1000 bootstrap replicates
#> and a sample of size: 100