Summary Bootstrap

This summary.bootstrap function is a method of the basic summary function. It will print a summary regarding the bootstrap class generated by the Bootstrap function. The returned values will be of class summary.bootstrap.

# S3 method for bootstrap
summary(x, level, digits, ...)

Arguments

x

is the output-list returned from the Bootstrap function

level

represent the confidence level of 1-alpha. By default level = 0.95

digits

are the decimals points desired in the printed output. By default digits = 2

...

other parameters for the 'summary.bootstrap' function

Value

A print of the main summary information from the Bootstrap function. It also return a list composed as follow:

summary

Will return a list composed of the basic summary function output applied on the theta_star output

CI

It will return a list with the lower and upper limit of the confidence interval

sto

The theta_hat from the Bootstrap function

coerced_to_a_vector

Returning a True or Flase regarding the initial sample if it was coerced to a vector during the usage of Bootstrap function

level

Represent the confidence level of 1-alpha called as a parameter

stat

Return the name of the statistic used thruough the Bootstrap function

B

Number of bootrsapped samples thourgh the Bootstrap function

Author

Group-E

Examples

boot <- Bootstrap(rnorm(100), statistic = var)
summary(boot, level = 0.90)
#> Non-parametric bootstrap for 'var'
#> Number of bootstrap replicates: 1000 
#> 
#> Statistic on the sample 'x': var(x)=0.78
#> 
#> Confidence interval at the 90%-level:
#> 0.62 0.95 
#> 
#> Bootstrap estimates:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    0.46    0.71    0.77    0.78    0.84    1.15